+++ CHANGELOG.md
@@ -5,6 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] - 2022-06-12
+
+### Added
+
+- 💅 General style: added "primary" and "secondary" colors for multiple elements.
+- 🔘 `Button`: add `useFocusStyle` to control if we'll use focus style directly
+- 📜 `Menu`: add more fine-grain control over rendering menu items
+
+### Changed
+
+- 🏊 `Toggle`, `Badge`, `RadioGroup`: `status` -> `colorType`
+- Internal: refactor color for different ColorType in scss
+- 👮 `Badge`: make `text` optional
+
+### Fixed
+
+- Various forms: Fix generic FormFields
+
 ## [1.2.6] - 2022-06-10
 
 ### Added